# Conflicts:
#	build.gradle
#	src/test/java/com/syntifi/casper/sdk/model/AbstractJsonTests.java
#	src/test/java/com/syntifi/casper/sdk/model/clvalue/encdec/EncoderDecoderTests.java
#	src/test/java/com/syntifi/casper/sdk/service/AbstractJsonRpcTests.java
#	src/test/java/com/syntifi/casper/sdk/service/CasperServiceTests.java

+++ src/main/java/com/syntifi/casper/sdk/model/transfer/Transfer.java
@@ -18,7 +18,7 @@ import lombok.Data;
 @Data
 public class Transfer {
     @JsonProperty("id")
-    private Long id;
+    private BigInteger id;
 
     /**
      * Hex-encoded account hash.
